Move patch handlers to separate files.
The patch handler contains logic for multiple request body types.
Moving them to separate files facilitates adding support for more types.
